pyproject.toml
[project]
name = "poc-ai-extraction"
version = "0.1.0"
description = "WT Daily Report AI Extraction POC"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "fastapi>=0.135.1",
    "openai>=2.26.0",
    "openpyxl>=3.1.5",       # Excel read engine used by pandas
    "pandas>=2.2.0",         # Excel parsing for .xlsx/.xls files
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
    "xlrd>=2.0.1",           # pandas engine for legacy .xls files (not just .xlsx)
]


